+++ ios/Flow Widgets/AppIntent.swift
@@ -10,9 +10,9 @@ import AppIntents
 
 struct ConfigurationAppIntent: WidgetConfigurationIntent {
     static var title: LocalizedStringResource { "Configuration" }
-    static var description: IntentDescription { "This is an example widget." }
+    static var description: IntentDescription { "Flow widget for entering transactions swiftly." }
 
     // An example configurable parameter.
-    @Parameter(title: "Favorite Emoji", default: "😃")
-    var favoriteEmoji: String
+//    @Parameter(title: "Favorite Emoji", default: "😃")
+//    var favoriteEmoji: String
 }